Fixed a small error when opening files
Previously, when opening new files, all the code in the current file would get deleted and cancelling the opening of a new file would throw an error. Now it doesn't do that anymore. Also, the name of the window no longer shows the filepath of the current file, it only shows it's name and file format.